[FSTORE-808] Statistics API - new schemas (#989)
* [FSTORE-808] Statistics API - new schemas

* Fix fds dto and class

* Fixes and improvements

* More fixes and java client

* Fix pytests

* Fix pytests after rebasing

* Address comments

+++ java/hsfs/src/main/java/com/logicalclocks/hsfs/metadata/RestDto.java
@@ -28,7 +28,7 @@ import java.util.List;
 public abstract class RestDto<D> {
 
   protected List<D> items = new ArrayList<>();
-  protected Integer count = null;
+  protected Long count = null;
 
   public RestDto() {
   }
@@ -41,11 +41,11 @@ public abstract class RestDto<D> {
     this.items = items;
   }
 
-  public void setCount(Integer count) {
+  public void setCount(Long count) {
     this.count = count;
   }
 
-  public Integer getCount() {
+  public Long getCount() {
     return count;
   }
